Touchtypegame

+++ GamePortal/GamePortal.Web.Api/Controllers/TouchType/TTGTextsController.cs
@@ -4,6 +4,9 @@ using System.Linq;
 using System.Net;
 using System.Net.Http;
 using System.Web.Http;
+using CSharpFunctionalExtensions;
+using FluentValidation;
+using JetBrains.Annotations;
 using Kbalan.TouchType.Logic.Dto;
 using Kbalan.TouchType.Logic.Services;
 
@@ -16,10 +19,12 @@ namespace GamePortal.Web.Api.Controllers.TouchType
     public class TTGTextsController : ApiController
     {
         private readonly ITextSetService _textSetService;
+        private readonly IValidator<TextSetDto> _textSetValidator;
 
-        public TTGTextsController(ITextSetService textSetService)
+        public TTGTextsController([NotNull]ITextSetService textSetService, [NotNull]IValidator<TextSetDto> TextSetValidator)
         {
             this._textSetService = textSetService;
+            _textSetValidator = TextSetValidator;
         }
 
         //Get All TextSets
@@ -27,7 +32,8 @@ namespace GamePortal.Web.Api.Controllers.TouchType
         [Route("")]
         public IHttpActionResult GetAll()
         {
-            return Ok(_textSetService.GetAll());
+            var result = _textSetService.GetAll();
+            return result.IsSuccess ? Ok(result.Value) : (IHttpActionResult)BadRequest(result.Error);
         }
 
         //Get TextSet by Id
@@ -35,7 +41,13 @@ namespace GamePortal.Web.Api.Controllers.TouchType
         [Route("{id}")]
         public IHttpActionResult GetById(int id)
         {
-            return _textSetService.GetById(id) == null ? (IHttpActionResult)NotFound() : Ok(_textSetService.GetById(id));
+            if (id <= 0)
+            {
+                return BadRequest("ID must be greater than 0");
+            }
+
+            var result = _textSetService.GetById(id);
+            return result.IsSuccess ? Ok(result.Value) : (IHttpActionResult)BadRequest(result.Error);
         }
 
         ///Get Random TextSet by Level of the text
@@ -43,7 +55,14 @@ namespace GamePortal.Web.Api.Controllers.TouchType
         [Route("searchbylevel/{level}")]
         public IHttpActionResult GetRandomByLevel(int level)
         {
-            return _textSetService.GetByLevel(level) == null ? (IHttpActionResult)NotFound() : Ok(_textSetService.GetByLevel(level));
+            if (level < 0 || level > 2)
+            {
+                return BadRequest("Level must be Easy, Middle or Hard");
+            }
+
+            var result = _textSetService.GetByLevel(level);
+            return result.IsSuccess ? Ok(result.Value) : (IHttpActionResult)BadRequest(result.Error);
+
         }
 
         //Add new text
@@ -51,16 +70,36 @@ namespace GamePortal.Web.Api.Controllers.TouchType
         [Route("")]
         public IHttpActionResult Add([FromBody]TextSetDto model)
         {
-            return _textSetService.Add(model) == null ? (IHttpActionResult)Conflict() : Created($"/textsets/{model.Id}", model);
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
+            var preValidResult = _textSetValidator.Validate(model, ruleSet: "PreValidation");
+            if (!preValidResult.IsValid)
+            {
+                return BadRequest(preValidResult.Errors.Select(x => x.ErrorMessage).First());
+            }
+
+            var result = _textSetService.Add(model);
+            return result.IsSuccess ? Created($"/textsets/{result.Value.Id}", result.Value) : (IHttpActionResult)BadRequest(result.Error); 
         }
 
-        //Update Text by Id
+        //Update Text by Id 
         [HttpPut]
         [Route("")]
         public IHttpActionResult Update([FromBody]TextSetDto model)
         {
-            _textSetService.Update(model);
-            return StatusCode(HttpStatusCode.NoContent);
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
+            var preValidResult = _textSetValidator.Validate(model, ruleSet: "PreValidation");
+            if (!preValidResult.IsValid)
+            {
+                return BadRequest(preValidResult.Errors.Select(x => x.ErrorMessage).First());
+            }
+
+            var result = _textSetService.Update(model);
+            return result.IsSuccess ? Ok($"Text set with id {model.Id} updated succesfully!") : (IHttpActionResult)BadRequest(result.Error);
+
         }
 
         //Delete Text by Id
@@ -68,8 +107,13 @@ namespace GamePortal.Web.Api.Controllers.TouchType
         [Route("{id}")]
         public IHttpActionResult Delete(int id)
         {
-            _textSetService.Delete(id);
-            return StatusCode(HttpStatusCode.NoContent);
+            if (id <= 0)
+            {
+                return BadRequest("ID must be greater than 0");
+            }
+
+            var result = _textSetService.Delete(id);
+            return result.IsSuccess ? Ok($"Text set with id {id} deleted succesfully!") : (IHttpActionResult)BadRequest(result.Error);
         }
     }
 }
